refactor(extraction): separate label and parameter into own modules
Further refine the extraction directory structure by separating label and
parameter parsing logic into their own dedicated files, matching the AST
elements pattern where annotations, labels, and parameters are distinct modules.

Structure changes:
- extraction/label.rs (74 lines) - Label token parsing logic
- extraction/parameter.rs (167 lines) - Parameter key=value parsing
- extraction/annotation.rs (92 lines) - Orchestrates label and parameter extraction

This mirrors the organization in ast/elements/ where label.rs and parameter.rs
are separate from annotation.rs, improving code clarity and maintainability.
